serverless/pytorch/centernet-mbv2-baiguang-29/nuclio/requirements.txt
numpy
opencv-python-headless
Pillow
torch==1.4.0
pyyaml
